refactor(ui): extract shared diff renderers to break the chat components import cycle
