+++ scripts/control-ui-mock-dev.ts
@@ -189,15 +189,15 @@ function buildSessionDiffMock() {
     "--- a/src/app.ts",
     "+++ b/src/app.ts",
     "@@ -12,4 +12,5 @@ export function bootstrap() {",
-    "   const config = loadConfig();",
+    "   const config = readSettings();",
     "-  const client = createClient(config);",
     "+  const client = createClient(config, { retries: 3 });",
-    "+  client.on(\"error\", reportError);",
+    '+  client.on("error", reportError);',
     "   return client;",
     "@@ -181,3 +182,3 @@ export function shutdown() {",
     "   flushQueues();",
-    "-  logger.info(\"bye\");",
-    "+  logger.info(\"shutdown complete\");",
+    '-  logger.info("bye");',
+    '+  logger.info("shutdown complete");',
     "",
   ].join("\n");
   const readmePatch = [
